Add timeout functionality.
I have added a timeout functionality. If the user does not
send a command within 30 seconds the socket will be destroyed.
I have also implemented an Executor Service and Callable duo for
authentication requests. This allows the main thread to continue
if the user has not authenticated after a certain amount of time.

Notifications also disapear after 30 seconds.